Real_folder/sort_jetson/requirements.txt
filterpy>=1.4.5         

gdown==5.0.0            # google drive model download
GitPython>=3.1.0        # track eval cloning
lapx>=0.5.4
loguru>=0.7.0
chardet>=5.2.0
numpy	                # otherwise issues with track eval
opencv-python>=4.6.0
pandas>=1.1.4           # export matrix
pre-commit>=3.3.3
PyYAML>=5.3.1           # read tracker configs
regex>=2023.6.3         # clip
ftfy>=6.1.1             # clip

scikit-learn>=1.3.0     # gsi
tensorboard>=2.13.0

# Export ----------------------------------------------------------------------

onnx>=1.12.0          # ONNX export
onnxsim>=0.4.1        # ONNX simplifier
nvidia-pyindex        # TensorRT export
nvidia-tensorrt       # TensorRT export
openvino-dev>=2022.3  # OpenVINO export
onnx2tf>=1.10.0       # TFLite export
